refactor: scrollRestore 로직 분리
